backend/requirements.txt
flask
requests
beautifulsoup4
dnspython
reportlab
python-nmap
urllib3
google-generativeai